+++ examples/PhasorDynamics/Tiny/ThreeBus/Basic/ThreeBusBasicJson.cpp
@@ -10,9 +10,7 @@
  */
 #include <cmath>
 #include <ctime>
-#include <filesystem>
-#include <fstream>
-#include <vector>
+#include <iostream>
 
 #include <GridKit/Model/PhasorDynamics/ComponentLibrary.hpp>
 #include <GridKit/Model/PhasorDynamics/SystemModel.hpp>
@@ -20,7 +18,7 @@
 #include <GridKit/Solver/Dynamic/Ida.hpp>
 #include <GridKit/Testing/TestHelpers.hpp>
 #include <GridKit/Testing/Testing.hpp>
-#include <GridKit/Utilities/CliOptions/CliOptions.hpp>
+#include <GridKit/Utilities/CliArgs/CliArgs.hpp>
 
 #define ERROR_TOL 1.0e-4
 
@@ -28,64 +26,47 @@ using scalar_type = double;
 using real_type   = double;
 using index_type  = size_t;
 
-// NOTES:
-// Write function to compare to CSV files
-//  compare number of lines and number of items per line
-
 int main(int argc, const char* argv[])
 {
   using namespace GridKit::PhasorDynamics;
   using namespace AnalysisManager::Sundials;
   using namespace GridKit::Utilities;
+  using namespace GridKit::Testing;
 
-  auto error_allowed = static_cast<real_type>(1e-4);
+  CliArgs args{
+      {
+          .name     = {"--case", "-c"},
+          .help     = "JSON file describing the system configuration",
+          .required = false,
+          .defaults = "ThreeBusBasic.case.json",
+      },
 
-  //
+      {
+          .name     = {"--compare", "-r"},
+          .help     = "Two CSV files to compare:\n"
+                      "<expected-output-file> <reference-file>",
+          .nargs    = 2,
+          .defaults = {"mon.csv", "ThreeBusBasic.ref.csv"},
+      },
+
+      {
+          .name     = {"--tolerance", "-t"},
+          .help     = "Allowable maximum error between compared files",
+          .type     = ArgType::Real,
+          .defaults = 1.0e-4,
+      }};
+
+  args.parseArgs(argc, argv);
+
   // Input file
-  //
-
-  std::filesystem::path input_file;
-  if (argc < 2)
-  {
-    if (std::filesystem::exists("ThreeBusBasic.case.json"))
-    {
-      input_file = std::filesystem::current_path() / "ThreeBusBasic.case.json";
-    }
-    else
-    {
-      std::cout << "\n"
-                   "ERROR: No input file found or provided.\n"
-                   "\n"
-                   "Usage:\n"
-                   "       ThreeBusBasicJson <json-input-file>\n"
-                   "\n"
-                   "Please provide a JSON input file as a positional command-line \n"
-                   "argument.\n"
-                   "\n"
-                   "By default this example will look for \"ThreeBusBasic.json\" in the \n"
-                   "current working directory and use that if found.\n"
-                   "\n";
-      exit(1);
-    }
-  }
-  else
-  {
-    input_file = argv[1];
-  }
-
+  auto input_file = args["case"]();
   std::cout << "Example: ThreeBusBasicJson\n";
   std::cout << "Input file: " << input_file << '\n';
 
-  //
   // Create model data
-  //
-
   auto data = parseSystemModelData(input_file);
 
-  //
   // Instantiate system
-  //
-
   SystemModel<scalar_type, index_type> sys(data);
   sys.allocate();
 
@@ -123,13 +104,24 @@ int main(int argc, const char* argv[])
   sys.stopMonitor();
 
   // Generate aggregate errors comparing variable output to reference solution
-  auto errorSet =
-      GridKit::Testing::compareCSV("mon.csv", "ThreeBusBasic.ref.csv");
+  TestStatus status{__func__};
+  if (!args["compare"].empty())
+  {
+    const auto& [out_file, ref_file] = args["compare"].as<2>();
 
-  // Print the errors
-  errorSet.display();
+    auto errorSet = GridKit::Testing::compareCSV(out_file, ref_file);
+
+    // Print the errors
+    errorSet.display();
+
+    auto error_allowed = args["tolerance"].as<real_type>();
+
+    status *= errorSet.total.max_error < error_allowed;
+
+    status.report();
+  }
 
   std::cout << "\n\nComplete in " << (stop - start) / CLOCKS_PER_SEC << " seconds\n";
 
-  return errorSet.total.max_error < error_allowed ? 0 : 1;
+  return status.get();
 }